+++ src/GRA.Controllers/ViewModel/MissionControl/Challenges/ChallengesDetailViewModel.cs
@@ -1,9 +1,9 @@
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.ComponentModel.DataAnnotations;
 using GRA.Domain.Model;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc.Rendering;
-﻿using System.Collections.Generic;
-using System.ComponentModel;
-using System.ComponentModel.DataAnnotations;
 
 namespace GRA.Controllers.ViewModel.MissionControl.Challenges
 {
