+++ src/GRA.Controllers/ViewModel/MissionControl/Triggers/TriggersDetailViewModel.cs
@@ -1,9 +1,9 @@
-using GRA.Domain.Model;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc.Rendering;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
+using GRA.Domain.Model;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc.Rendering;
 
 namespace GRA.Controllers.ViewModel.MissionControl.Triggers
 {
