+++ src/GRA.Data/Repository/SchoolRepository.cs
@@ -20,7 +20,7 @@ namespace GRA.Data.Repository
         }
 
         public async Task<ICollection<School>> GetAllAsync(int siteId,
-            int? districtId = default(int?))
+            int? districtId = default)
         {
             var schoolList = DbSet
                 .AsNoTracking()
@@ -89,6 +89,21 @@ namespace GRA.Data.Repository
                 .AsNoTracking()
                 .Where(_ => _.Id == schoolId && _.SiteId == siteId)
                 .AnyAsync();
+        }
+
+        public async Task<IList<SchoolImportExport>> GetForExportAsync()
+        {
+            return await DbSet
+                .Include(_ => _.SchoolDistrict)
+                .OrderBy(_ => _.SchoolDistrict.Name)
+                .ThenBy(_ => _.Name)
+                .AsNoTracking()
+                .Select(_ => new SchoolImportExport
+                {
+                    District = _.SchoolDistrict.Name,
+                    Name = _.Name
+                })
+                .ToListAsync();
         }
     }
 }
