+++ src/GRA.Domain.Repository/ISchoolRepository.cs
@@ -7,7 +7,7 @@ namespace GRA.Domain.Repository
 {
     public interface ISchoolRepository : IRepository<School>
     {
-        Task<ICollection<School>> GetAllAsync(int siteId, int? districtId = default(int?));
+        Task<ICollection<School>> GetAllAsync(int siteId, int? districtId = default);
 
         Task<bool> IsInUseAsync(int siteId, int schoolId);
 
@@ -15,5 +15,6 @@ namespace GRA.Domain.Repository
         Task<int> CountAsync(BaseFilter filter);
 
         Task<bool> ValidateAsync(int schoolId, int siteId);
+        Task<IList<SchoolImportExport>> GetForExportAsync();
     }
 }
