+++ src/GRA.Web/Areas/MissionControl/Views/Schools/Districts.cshtml
@@ -52,7 +52,7 @@
                                             data-target="#editModal"
                                             data-id="@district.Id"
                                             data-name="@district.Name">
-                                        <span class="fa fa-2x fa-pencil" aria-hidden="true"></span>
+                                        <span class="fa fa-pencil" aria-hidden="true"></span>
                                     </button>
                                     <button type="button"
                                             class="btn btn-danger btn-xs"
@@ -61,7 +61,7 @@
                                             data-target="#deleteModal"
                                             data-id="@district.Id"
                                             data-name="@district.Name">
-                                        <span class="fa fa-2x fa-remove" aria-hidden="true"></span>
+                                        <span class="fa fa-remove" aria-hidden="true"></span>
                                     </button>
                                 </td>
                             </tr>
