+++ src/GRA.Web/Areas/MissionControl/Views/Schools/_SchoolsPartial.cshtml
@@ -3,6 +3,20 @@
         <ul class="nav nav-pills" style="padding: 10px 0;">
             <li ActiveBy routeKey="Action" value="Index"><a asp-action="Index">Schools</a></li>
             <li ActiveBy routeKey="Action" value="Districts"><a asp-action="Districts">School Districts</a></li>
+            <li class="pull-right">
+                <div>
+                    <a asp-action="Import"
+                       ActiveBy routeKey="Action" value="Import"
+                       class="btn btn-sm btn-default">
+                        <span class="fas fa-file-upload"></span>
+                        Import
+                    </a>
+                    <a asp-action="Export" class="btn btn-sm btn-default">
+                        <span class="fas fa-file-download"></span>
+                        Export
+                    </a>
+                </div>
+            </li>
         </ul>
     </div>
 </div>
