+++ src/GRA.Web/Areas/MissionControl/Views/Systems/Index.cshtml
@@ -1,6 +1,5 @@
 ﻿@model GRA.Controllers.ViewModel.MissionControl.Systems.SystemListViewModel
 
-
 <div class="row">
     <div class="col-xs-12">
         <ul class="nav nav-pills" style="padding: 10px 0;">
